test Compiler.implicit_rpaths
