+++ lib/spack/spack/test/llnl/util/filesystem.py
@@ -9,7 +9,8 @@ import llnl.util.filesystem as fs
 import os
 import stat
 import pytest
-import itertools
+
+from spack.test.conftest import dirs_with_libfiles
 
 
 @pytest.fixture()
@@ -196,31 +197,6 @@ class TestInstallTree:
 
             assert os.path.exists('dest/2')
             assert not os.path.islink('dest/2')
-
-
-@pytest.fixture()
-def dirs_with_libfiles(tmpdir_factory):
-    lib_to_libfiles = {
-        'libstdc++': ['libstdc++.so', 'libstdc++.tbd'],
-        'libgfortran': ['libgfortran.a', 'libgfortran.dylib'],
-        'libirc': ['libirc.a', 'libirc.so']
-    }
-
-    root = tmpdir_factory.mktemp('root')
-    lib_to_dirs = {}
-    i = 0
-    for lib, libfiles in lib_to_libfiles.items():
-        dirs = []
-        for libfile in libfiles:
-            root.ensure(str(i), dir=True)
-            root.join(str(i)).ensure(libfile)
-            dirs.append(str(root.join(str(i))))
-            i += 1
-        lib_to_dirs[lib] = dirs
-
-    all_dirs = list(itertools.chain.from_iterable(lib_to_dirs.values()))
-
-    yield lib_to_dirs, all_dirs
 
 
 def test_paths_containing_libs(dirs_with_libfiles):
